requirements.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
grpcio==1.75.1
h5py==3.14.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
keras==3.11.3
keras-facenet==0.3.2
kiwisolver==1.4.9
lz4==4.4.4
markdown-it-py==4.0.0
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mtcnn==1.0.0
numpy==1.26.0
opencv-contrib-python==4.12.0.88
opt_einsum==3.4.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.1
pyarrow==10.0.1
pyparsing==3.2.5
pytz==2025.2
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.2
scipy==1.15.3
streamlit==1.50.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.3
